+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Introduce the concept of the `FATAL_ERROR` state for connectors where no attempts to reconnect accor. This state is set when we get an error message that starts with the string `"java.lang.SecurityException:"` under the assumption that a user will not recover until given appropriate permissions.
+
 ### [v6.128](https://github.com/replicant4j/replicant/tree/v6.128) (2022-10-17) · [Full Changelog](https://github.com/spritz/spritz/compare/v6.127...v6.128)
 
 Changes in this release: